requirements.txt
python-dotenv
numpy
openai
flask
flask-cors
